[Refactor] AppDetailInfoView 하위 뷰로 분리
- 앱 디테일 정보 영역 하위 뷰로 분리
